[FIX] - Fix code task 6

+++ 0x00-hello_world/6-size.c
@@ -11,7 +11,7 @@ int intT;
 float floatT;
 char charT;
 long int longIT;
-long int longDT;
+long long int longDT;
 
 printf("Size of a char: %ld byte(s)\n", sizeof(charT));
 printf("Size of an int: %ld byte(s)\n", sizeof(intT));
